Fork LegacyHydrawiseAsync to rest.RestClient
